+++ NEWS.md
@@ -5,6 +5,10 @@
 - `ensemble_nested_weighted()`: Apply weighted ensembles iteratively
 - `ensemble_nested_average()`: Apply average ensembles iteratively
 
+## New Vignette
+
+- Nested Ensembles
+
 # modeltime.ensemble 0.4.2
 
 ### Compatibility with `modeltime` 0.7.0.